Manoj
Author: srivasta
Date: 2002/11/15 06:44:22
   Manoj
* Added example for why one may call ldconfig anywhere in the
postsint.                                                closes: Bug#120585
* Add the modifications about base system, as opposed to the soon to be
obsolete base section (I assume it is)                    closes: Bug#53582
* Rearranged the virtual packages list.                     closes: Bug#72980
* This is basically an attempt to ratify the current practice of using
debhelper in the clean target. Currently, policy does not require
debhelper to be installed when the clean target is run, even though it
is in the build-depends field. This was a simple oversight.
closes: Bug#164035
* No longer depend on fileutils.                           closes: Bug#167425
* Added the Apps/Science menu section                      closes: Bug#162812
* Applied text patch from Joey Hess to the debconf spec simply make it
conform to the reality of how some things work now. This is part of an
effort to make debconf and cdebconf better substitutes for each
other. Since it was not an XML patch, no special markup is present in
the new content, except where I made guesses.            closes: Bug#160776
* Clarify section 13.3.                                    closes: Bug#160248
* Removed the undocumented(7) hack requirement.            closes: Bug#39830
Josip
* Removed the obsolete notion of documenting changes within the copyright
file,                                                     closes: Bug#65764

git-archimport-id: srivasta@debian.org--etch/debian-policy--devel--3.0--patch-156