Debian changelogs should be UTF-8 encoded. Changed the wording from a
Author: srivasta
Date: 2003/06/07 17:39:17
Debian changelogs should be UTF-8 encoded. Changed the wording from a
should to a may; since a should would make an unknown number of
packages insta buggy. A reuest makes all these wishlist bugs; we can
raise the severity in a later version of policy.

Added LANG=C before the debiandoc2X invocations, this ensures that the
resulting documents do not get converted to the locale on the building
machine.

git-archimport-id: srivasta@debian.org--etch/debian-policy--devel--3.0--patch-230