Added an instance if what recovery is attempted if error unwind itself
Author: srivasta
Date: 2005/02/05 00:37:19
Added an instance if what recovery is attempted if error unwind itself
runs into problems.

Cleaned up some of the language in the /etc/init.d script output
formatting.

Cleaned up the command -v example. Instead of the massively complex
shenanigans suggested in the bug report, just use which. which lives
in an essential package, so can be used in the preinst of packages.

git-archimport-id: srivasta@debian.org--etch/debian-policy--devel--3.0--patch-280