Add a recommendation for the location for run-time support material,
Author: srivasta
Date: 2005/02/05 01:48:49
Add a recommendation for the location for run-time support material,
not meant to be invoked directly.
a) it shows a predilection for a directory named for the package,
but does not mandate it
b) Since this is a new requirement, it starts off as a
recommendation (after all, this was originally touted merely as
"best practice", not as something required for interoperability)
c) It also discusses arch independent modules, and adds similar
language for that.

git-archimport-id: srivasta@debian.org--etch/debian-policy--devel--3.0--patch-283