Add test -a/-o and local to shell policy, update wording
Scripts may now assume that /bin/sh supports local (in a limited form)
and, if they use a built-in test, that built-in supports test -a/-o
binary logical operators.  Reword to include echo -n as another case
of additional required features.  Replace references to POSIX with
SUSv3 and add a link.

git-archimport-id: rra@debian.org--lenny/debian-policy--devel--3.7--patch-2