Document the type field in shlibs and its use with udebs
The shlibs file syntax now supports an optional type field specifying
the type of packages to which the line applies.  Currently, the only
supported type is udeb.  Include documentation of the syntax, the use
with udeb, and an example.  Based on a patch from Frans Pop with input
from Raphael Hertzog.

git-archimport-id: rra@debian.org--lenny/debian-policy--devel--3.7--patch-15