Standardize manual page encoding
Standardize manual page encoding to either the legacy encoding for a
locale or UTF-8.  Country names should not be included in locale-specific
directory names unless there is a significant difference in language.
Characters in the manual page must be representable in the legacy encoding
for now.

Patch from Colin Watson.  Closes #440420.

git-archimport-id: rra@debian.org--lenny/debian-policy--devel--3.7--patch-31