Add README.source and the patch target
Specifies the debian/README.source file to document any package that
uses a source or package management system or that has a complex source
layout or build system.  Specifies and recommends the patch target
for putting a source package into an editable state.  Allowed referring
to external documentation in README.source if it's specific and not a
general manual.

git-archimport-id: rra@debian.org--lenny/debian-policy--devel--3.7--patch-41