Clarify support file handling for shared libraries
Combine the discussion of run-time support files for shared libraries
in a renamed section.  Be explicit that shared library packages may not
contain any file whose name does not change when the soname changes,
and be clearer about when files should be put into versioned directories
in the shared library package and when files should go into a separate
package.  Files useful only when compiling against the library should
go into the -devel package (and other run-time support programs should
not).

Suggest -tools instead of -runtime for the package containing support
files or binaries built against a library and usable by any version of
the library.  Thanks, Raphael Hertzog.