Policy: minor typos in dpkg-shlibdeps and dpkg-architecture
Fix minor typos in appendix C in the sections on dpkg-shlibdeps and
dpkg-architecture.  Also remove the note that future versions of
dpkg-shlibdeps will be called on shared libraries as well and just
say that it acts on both executables and shared libraries.

Thanks to Sandro Tosi <matrixhasu@gmail.com> for the corrections.
Closes: #481954