Clarify handling of diversions in maintainer scripts
Suggest running dpkg-divert unconditionally in the preinst to also
catch cases of upgrading a package from a version before the diversion
was added.  Suggest an alternate form that doesn't re-add the diversion
when unnecessary on upgrades.

In the postrm example, remove the diversion in more cases and document
why it should not be removed on upgrades.

Closes: #483418