Also handle the abort-upgrade postrm case
When the diversion was added in a particular version of a package,
an aborted upgrade to a newer version with the diversion from an
older version without the diversion should remove the diversion when
rolling back to the older version.  Add the necessary postrm code and
explanation.  Thanks to Raphael Hertzog for the review.