Allow the Binary control field to wrap
Allow the Binary control field to wrap in both *.dsc and *.changes files,
since we have some source packages that produce lines so long that GnuPG
is no longer willing to sign the file.

Clearly state that the syntax is different depending on whether the field
appears in a *.dsc file or a *.changes file.

Thanks to Raphaël Hertzog for the report.  Debian Bug#533852.