[bug545548-srivasta]: Add Documentation
This patch add a README file, rendered as txt and html, and also
documents the policy change process, again rendered as text and
HTML. While the text and HTML files are automatically generated, they are
shipped in the package itself so as to avoud depending on a recent
version of Emacs during build.

The contents of the new documents are based on (but no identical to) the
contents of related pages on the Debian wiki. The long term plan is to
make these documents the canonical ones, and have the Wiki point to these
pages.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>