[master]: Added changelog for bug 545548
This patch add a README file, rendered as txt and html, and also
documents the policy change process, again rendered as text and
HTML. While the text and HTML files are automatically generated, they are
shipped in the package itself so as to avoud depending on a recent
version of Emacs during build.  If a new Emacs is present, the .txt and
.html files shall be regenerated if the org file is newer.

The contents of the new documents are based on (but no identical to) the
contents of related pages on the Debian wiki. The long term plan is to
make these documents the canonical ones, and have the Wiki point to these
pages. Also render out a upgrading-checklist.txt (we can also do a
docbook version later)

Closes: #545548

Signed-off-by: Manoj Srivastava <srivasta@debian.org>