Specification of date format -- bug #569174.
RFC 822 dates use only two digits for the years, but Debian changelogs
described by this paragraph (§4.4 in Policy 3.8.4) use four digits.

This patch replaces the reference to the RFC 822 by a specification that is
compatible with its successors, RFC 2822 and RFC 5322, but does not use their
full range of options.