Tighten requirements for maintainer-like fields
The syntax of Maintainer is a must, not a should.  The comma
separator in Uploaders is a must, not a should.  Now allow people
to use multi-line Uploaders fields freely.  Be explicit that the
Changed-By field has the same syntax as Maintainer and be clearer
about what identity goes here.