Significantly rework the discussion of Breaks and Conflicts
Recommend Breaks over Conflicts for most purposes.  List the places
Breaks should be used and the places Conflicts should be used.  Add
Breaks to the example in Replaces, and add a footnote explaining
why Breaks is normally required along with Replaces when files are
moving between packages.  Add additional discussion of conflicting
with a virtual package.