Further revisions to the Format update
Drop historical information about the version changes in .changes
files.  It doesn't seem useful at this point.  Specify the format
of Format for source packages separately since referring to the
version format with exceptions and additions is more confusing and
it's not really a version in the same sense.  Add the full list of
supported source package formats in a footnote.  Add a bit more
explanation for how the source package format is used.