Rewrite requirements around libtool .la file handling
State that .la files should normally not be included in Debian
packages, with special exceptions for libraries loaded by libltdl and
for .la files on which other .la files may depend.  State that
dependency_libs should be emptied if the .la files have to be
retained.  Exclude non-public libraries and loadable modules from
these requirements.