Clarify Maintainer, make Uploaders mandatory for team maintainers
Clarify the duties of the maintainer and state that the maintainer's
e-mail address must accept non-spam BTS mail and automated mail from
the project software.  Rephrase the package orphaning; this does not
only happen when someone retires from the project.  Require the
Uploaders field for a team-maintained package and, in that case,
require that it contains at least one human and personal address.