Fix multiarch FHS exception for i386 in light of recent discussions
The current value of DEB_HOST_GNU_TYPE on i386 is unsuitable for
cross-distro standardization, because it varies according to the default CPU
target of the toolchain.  Discussion with the toolchain and dpkg maintainers
yielded an alternative solution, a new dpkg-architecture variable
DEB_HOST_MULTIARCH which is committed to dpkg upstream in commit
af3153d09aa3ed5597d6d415e5ab7cc3ba972e7c and will be included in the upload
of dpkg 1.16.0.  Update Policy to document this new requirement for
multiarch.