Architecture in .dsc may now be "any all"
With dpkg >= 1.16.1, the value of the Architecture field in the .dsc
is generated a bit differently.

The goal is to be able to always know when a source package generates an
architecture-independent package. Currently as soon as one binary is
"Architecture: any", the value generated is "any" and we have no way to
know that there are "Architecture: all".

With dpkg >= 1.16.1, the "all" is kept even if there's a binary package
with "any". So the resulting value is then "any all". This allows
wanna-build to know if it must schedule a build for the "Architecure: all"
packages.