Do not require initial Debian maintainers be listed in copyright
The documentation of debian/copyright explicitly required the
initial Debian maintainers be listed in the file.  That information
is already in the changelog, and they may no longer hold any
copyrights on the remaining files in the package.  Drop that
requirement, although the requirement to include copyright and
license information is still in force.

* Policy: Initial Debian maintainers need not be listed in copyright
  Wording: Charles Plessy <plessy@debian.org>
  Seconded: Russ Allbery <rra@debian.org>
  Seconded: Raphaël Hertzog <hertzog@debian.org>
  Seconded: Andrew McMillan <andrew@morphoss.com>
  Closes: #593533