Significant rewording of /run, add a separate section
Lift more of the properties of /run into normative text and state
explicitly that the FHS requirements apply to /run and /run/lock.
Add a new section explaining that packages cannot rely on structure
in /run unless they've created it and saying explicitly that packages
must not ship files in /run or /var/run or /var/lock.  Reference
that section from the init.d section.