DEP-5 updates from further debian-project discussion
Clarify that this document does not supersede or modify the
requirements in Policy and that use of the specification is
optional.  Note that a single Files paragraph can apply to the
whole package (rather than using the "all files" wording) and
that only the license and copyright information required by
the Debian archive has to be listed.

Clarify the wording of stand-alone license paragraphs to be
clear that they can be used with exceptions.

Some wording improvements, including making the attributes
listed after the fields lowercase.  Fix a few links and the
official keyword of the BSD-3-clause license in some examples.

Add a Version 1.0 subtitle.