General editing pass for the copyright format specification
Add a pointer to DEP-5 in the synopsis and change the verb tense
to present from imperative.  Rephrase the introduction to talk about
this document as a specification rather than a proposal.

Explicitly mention Files: * as the way to write a single files
paragraph that covers the entire package.

Mention after one of the examples that the two paragraphs for files
with the same license but different copyright holders could be combined
at the discretion of the document author.

Give the complete correct Format field for the current version of the
document in the description of that field.

Add more explanation for how Disclaimer is used.

Say explicitly that more general files paragraphs should be given
first, followed by more specific overrides.

Update the license list to note that "Universal" is removed from the
version of CC0 when forming the short name, to list the new versions
of the GFDL, and to be more explicit about when to use GFDL-NIV.

Rewording and rephrasing in various places to hopefully make the
text flow better, read more smoothly, and be more idiomatic.

Add multiple additional in-line examples, particularly around the
Files patterns and the formation of license short names.