Install HTML upgrading-checklist in policy.html directory
* Install the HTML version of upgrading-checklist in the policy.html
  directory as upgrading-checklist.html so that it can be deployed on
  www.debian.org in a way that will allow links to Policy sections to
  work easily.  Thanks, Charles Plessy.  (Closes: #639663)

For right now, this is done with a bit of a hack to copy the file into
the appropriate directory and to turn the one-page HTML version of it
into a symlink inside the constructed package.

The clean solution will be to make the upgrading checklist an appendix
of Policy, which may be done as part of a DocBook conversion.