Substantial revision to the new symbols draft
Pull out lots of information that's shared between symbols and shlibs
into new sections directly under shared library dependencies.  Make
both symbols and shlibs subsections of that section.  Flesh out the
documentation of how to choose between symbols and shlibs and what the
differences in the systems are.

Document that C++ libraries may be better off with shlibs and don't
deprecate shlibs quite as much.

Move the information on how to determine the shared library soversion
to the runtime shared library section where it's now primarily used,
and reference it in the shlibs section.