Further symbols wording fixes from Charles Plessy
Retain the dpkg-shlibdeps section in the appendix as a stub pointing
to the main page so that we don't renumber subsequent sections.  Add
more explanation for why dependencies change for each build of a
binary package.  Combine the dpkg-shlibdeps instructions somewhat and
mention that debian/substvars is only a default.